four files js to ts, fixed ng_model_on_blur
